validation added to session step
